requirements.txt
asgiref==3.8.1
certifi==2025.4.26
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
Django==5.2.3
django-channels==0.7.0
djangorestframework==3.16.0
idna==3.10
msgpack==1.1.1
oauthlib==3.2.2
redis==6.2.0
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
